photconv handles SB, absolute mags explicitly
Split the ab_to_mjy conversion function to three cases: converting a
magnitude to mjy, converting a surface brightness to mjy, and converting
an absolute magnitude to mjy. Previously only the latter case was
handled since we were working with mock FSPS tests. Having the other
cases makes it easier to work with observations.